Constrain the sensor holder instead of using AllMotion immobility
It turns out that this prevented the physbone from responding to the rotation
of its parent. A VRCPositionConstraint is now used to prevent it from moving
relative to the avatar root.

A single VRCParentConstraint could've been used to pin it to the world origin,
but I don't know if there are any consequences to that -- it could move the
physbone very far away.